+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Fixed `TabProvider` & `ProgressTrackerProvider` not calling `onChange` when the index is not provided via the props.
 -   Fixed `Link` default style when in button mode (padding + text align).
 
+### Added
+
+-   New component in React: `Flag`
+
 ## [1.0.8][] - 2021-02-08
 
 ### Added